+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.54.0] - Unpublished
+
+### Added
+
+- More bcs Optional types to support argument inferencing
+- Dataclass for Staked coins
+- BCS `Optional`` factory for GraphQL transaction builder
+- ObjectRef (ImmOrOwnedObject, SharedObject) added ingestion for ObjectReadGQL
+
+### Fixed
+
+### Changed
+
+- Bump Sui version support to include 1.18.x - 1.21.0
+
+### Removed
+
 ## [0.53.0] - 2024-02-29
 
 ### Added